requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community==0.0.38
google-generativeai==0.2.2
redis==5.0.0
celery==5.3.2
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.70
eventlet==0.33.3
python-socketio>=5.9.0
SQLAlchemy==2.0.21
Alembic>=1.12.0
gunicorn>=21.0.0
twilio==8.8.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dateutil==2.8.2
